// apps/frontend/lib/api/generated/models/CredibilityHighlight.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CredibilityHighlight = {
	quote?: string;
	sentiment?: CredibilityHighlight.sentiment;
};
export namespace CredibilityHighlight {
	export enum sentiment {
		POSITIVE = "POSITIVE",
		NEUTRAL = "NEUTRAL",
	}
}